+++ b_lambda_layer_common/util/pynamodb_encoder.py
@@ -5,6 +5,7 @@ from pynamodb.attributes import MapAttribute
 from b_lambda_layer_common.util.dynamodb_encoder import DynamoDBEncoder
 
 
+# TODO move to pynamo_db directory.
 class PynamoDBEncoder(DynamoDBEncoder):
     def default(self, o: Any):
         if isinstance(o, MapAttribute):