+++ b_lambda_layer_common_test/unit/pynamo_db/test_kms_attribute.py
@@ -1,0 +1,21 @@
+from b_lambda_layer_common.pynamo_db.kms_attribute import KmsAttribute
+
+
+def test_FUNC_serialize_WITH_null_value_EXPECT_serialzied() -> None:
+    """
+    Test whether the serialization works.
+
+    :return: No return.
+    """
+    attribute = KmsAttribute()
+    attribute.serialize(None)
+
+
+def test_FUNC_deserialize_WITH_null_value_EXPECT_deserialzied() -> None:
+    """
+    Test whether the serialization works.
+
+    :return: No return.
+    """
+    attribute = KmsAttribute()
+    attribute.deserialize(None)